Updates for the year 2018
Went through the document to update the surevey year, and added some clarification on which variables are for both years, and which aren't.

Still missing is a new scope for the read and write acces. For an user to show their score on their application. This may lead to an additional section in the api-doc.